requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
cryptography==42.0.8
httpx==0.27.0
python-dotenv==1.0.1
itsdangerous==2.2.0
